pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tfm-eda-agent"
version = "0.1.0"
description = "Agente de IA para automatización de EDA con PydanticAI + MCP + Ollama"
requires-python = ">=3.11"
dependencies = [
  "pydantic==2.12.5",
  "pydantic-ai-slim[mcp,openai]==1.70.0",
  "fastmcp==2.14.5",
  "pandas==2.2.3",
  "python-dotenv>=1.1.0",
  "rich==13.9.4",
  "streamlit>=1.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
